Review deps.edn (#85)
* Review deps.edn

CI setup-clojure action
- moved action to current release instead of @master
- renamed deprecated `tools-deps` input to `cli`
- move to clojure version 'latest' (just like we do for each major JDK version)

CI
- report tools' versions to allow for reproducibility

clj-watson simple safe-looking bumps:
- clojure
- edamame
- cheshire
- clj-http
- tools.deps (has not changed much)
- selmer

clj-watson less safe-looking:
- org.apache.maven.resolver/maven-resolver-transport-http (deleted) -
this dep is automatically included by tools.deps, and clj-watson does
not reference it directly, and I don't see a rationale in git history
for why we need to override the transitive version. It might have been a
CVE or bug sometime in the past, let's see if we still actually need it.

Dev only:
- nrepl
- clojure-lsp-standalone
  - also suppressed unused public var for scan entrypoint
  - committed new .clj-kondo lib exports
- kaocha
- test-check (deleted, unused)
- state-flow (deleted, unused)
- mockfn (delete, unused)

Other deps.edn changes:
- delete `:mvn/repos` entry as it is no different from defaults

Closes #75

* ci: review feedback: setup-clojure use major only

Specifying only the major version of the setup-clojure action is consistent
with how we specify the version of other actions.

And there will less bump maintenance, as we'll automatically use the
latest 12.x.

* ci: revert change: use 12.5 for setup-clojure

I was fully expecting @12 would resolve to the latest version of 12.x.
But it does not, so back to @12.5.